lk and kernel: localgit, mdm9x50 sierra files handling.
* common/classes/localgit.bbclass (do_fetch):
In Yocto 2.2, if we override the do_fetch task, the
fetching of SRC_URI material is suppressed (which is what you would
expect, but it didn't happen under Yocto 1.7). We were relying
on our custom do_fetch to create the git repo symlink, but then
SRC_URI material to still be fetched into the linked repo.
This is what fetches the "sierra files" kernel headers.
Under Yocto 2.2, we make this a _prepend action to achieve
the same thing. But we have to now make sure the SRC_URI is
empty in the base recipe.
(do_unpack, do_patch): We must remove these empty overrides.
In Yocto 2.2, it is do_unpack which actually propagates the
SRC_URI material into the target tree! do_fetch fetches the
materials, but the materials do not appear where they are
supposed to if do_unpack is suppressed. Under Yocto 1.7,
this wasn't the case.

* meta-swi-mdm9x50/recipes-kernel/linux/linux-quic_git.bb (SRC_URI):
Variable removed. A "localgit" procjet shouldn't specify the source
directory to be fetched; this is done via the SRC_DIR in the immediately
next line.
(do_unpack_append): This has to be removed, because do_unpack
is now Python code, and this is shell. The logic is still
needed, however, and is moved into do_configure_prepend,
where it appears in a modified form.
(do_configure_prepend): We create the ${STAGING_KERNEL_DIR} symlink here
instead of do_unpack_append. The symlink is a different: instead of
going to the source directory ${S} which is itself a symlink to the git
repo (thanks to localgit), it goes directly to the git repo. Moreover,
the symlink is not created unconditionally. Turns out, in Yocto 2.2,
Poky's kernel.bbclass recipe now creates this same symlink for us (as
an append action to do_unpack, too, just like what we achieved with our
do_unpack_append)! Unfortunately, this kernel.bbclass symlink creation
only occurs if SRC_URI is not empty. That is currently only the case on
mdm9x50 where we are using SRC_URI to bring in the sierra kernel
headers. Thus, we still need our own logic, in which we check whether
the symlink exists. We could make the link unconditionally in the 9x40
and 9x28 recipes, and not do it in the 9x50 recipe, but let's keep them
all the same in every possible regard.

* meta-swi-mdm9x28/recipes-kernel/linux/linux-quic_git.bb: Same change
as in mdm9x50.

* meta-swi-mdm9x40/recipes-kernel/linux/linux-quic_git.bb: Likewise.

Resolves: LXSWIREF-157

+++ meta-swi-mdm9x28/recipes-kernel/linux/linux-quic_git.bb
@@ -10,7 +10,6 @@ KERNEL_DEFCONFIG ?= "mdm9607_defconfig"
 B = "${WORKDIR}/build"
 KERNEL_EXTRA_ARGS        += "O=${B}"
 
-SRC_URI = "file://${LINUX_REPO_DIR}/../"
 SRC_DIR = "${LINUX_REPO_DIR}/.."
 
 LINUX_VERSION ?= "3.18.20"
@@ -19,20 +18,20 @@ PR = "r1"
 
 do_deploy[depends] += "dtbtool-native:do_populate_sysroot mkbootimg-native:do_populate_sysroot"
 
-do_unpack_append() {
-    # We do the following for the sake of other recipes which require
-    # ${STAGING_KERNEL_DIR} to hold the kernel source tree. Because our
-    # "localgit" based recipe bypasses much of the Yocto-supplied kernel class
-    # (such as the unpack steps), Yocto ends up with STAGING_KERNEL_DIR
-    # pointing to an empty directory. Recipes which refer to it will break.
-    # Example: our embms-kernel package. Solution: replace the empty dir with a
-    # symlink to our kernel tree.
-    mkdir -p ${STAGING_KERNEL_DIR}
-    rm -rf ${STAGING_KERNEL_DIR}
-    ln -sf ${S} ${STAGING_KERNEL_DIR}
-}
-
 do_configure_prepend() {
+    # When SRC_URI contains something, the Yocto kernel.bbclass creates
+    # ${STAGING_KERNEL_DIR} as a symlink to the local git repo; due to similar
+    # reasons that also motivate our localgit class. This is in the do_unpack
+    # step. When SRC_URI is empty, ${STAGING_KERNEL_DIR} ends up an empty
+    # directory rather than a symlink. We check for this and create the
+    # symlink anyway.
+    if [ ! -L ${STAGING_KERNEL_DIR} ] ; then
+      rm -rf ${STAGING_KERNEL_DIR}
+      mkdir -p ${STAGING_KERNEL_DIR}
+      rmdir ${STAGING_KERNEL_DIR}
+      ln -sf ${SRC_DIR} ${STAGING_KERNEL_DIR}
+    fi
+
     cp ${S}/arch/arm/configs/${KERNEL_DEFCONFIG} ${WORKDIR}/defconfig
 
     oe_runmake_call -C ${S} ${KERNEL_EXTRA_ARGS} mrproper
